+++ src/main/java/be/baur/sds/content/DateType.java
@@ -1,5 +1,6 @@
 package be.baur.sds.content;
 
+import be.baur.sds.common.Content;
 import be.baur.sds.common.Date;
 
 /**