+++ src/main/java/be/baur/sds/content/IntegerType.java
@@ -1,5 +1,7 @@
 package be.baur.sds.content;
 
+import be.baur.sds.common.Content;
+
 /**
  * A <code>IntegerType</code> represents an SDA node with integer content.<br>
  * See also {@link DecimalType}.