+++ src/main/java/be/baur/sds/serialization/SDSParser.java
@@ -13,6 +13,7 @@ import be.baur.sds.Component;
 import be.baur.sds.MixedType;
 import be.baur.sds.NodeType;
 import be.baur.sds.Schema;
+import be.baur.sds.common.Content;
 import be.baur.sds.common.Date;
 import be.baur.sds.common.DateTime;
 import be.baur.sds.common.Interval;
@@ -21,7 +22,6 @@ import be.baur.sds.content.AbstractStringType;
 import be.baur.sds.content.AnyType;
 import be.baur.sds.content.BinaryType;
 import be.baur.sds.content.BooleanType;
-import be.baur.sds.content.Content;
 import be.baur.sds.content.DateTimeType;
 import be.baur.sds.content.DateType;
 import be.baur.sds.content.DecimalType;
@@ -406,17 +406,17 @@ public final class SDSParser implements Parser {
 			return new NodeType(name); // remaining code does not apply
 		}
 		
-		MixedType mxType;	// the type returned at the end of this method
+		MixedType mixedType;	// the type returned at the end of this method
 		
 		switch (content) {
-			case STRING   : mxType = new StringType(name); break;
-			case BINARY   : mxType = new BinaryType(name); break;
-			case BOOLEAN  : mxType = new BooleanType(name); break;
-			case INTEGER  : mxType = new IntegerType(name); break;
-			case DECIMAL  : mxType = new DecimalType(name); break;
-			case DATETIME : mxType = new DateTimeType(name); break;
-			case DATE     : mxType = new DateType(name); break;
-			case ANY      : mxType = new AnyType(name); break;
+			case STRING   : mixedType = new StringType(name); break;
+			case BINARY   : mixedType = new BinaryType(name); break;
+			case BOOLEAN  : mixedType = new BooleanType(name); break;
+			case INTEGER  : mixedType = new IntegerType(name); break;
+			case DECIMAL  : mixedType = new DecimalType(name); break;
+			case DATETIME : mixedType = new DateTimeType(name); break;
+			case DATE     : mixedType = new DateType(name); break;
+			case ANY      : mixedType = new AnyType(name); break;
 			default: // will never get here, unless we forgot to implement something...
 				throw new RuntimeException("SDS type '" + content + "' not implemented!");
 		}	
@@ -426,8 +426,8 @@ public final class SDSParser implements Parser {
 		// Set the null-ability (not allowed on the any type).
 		Node nullable = getAttribute(sds, Attribute.NULLABLE, isAnyType? null : false);
 		if (nullable != null) switch(nullable.getValue()) {
-			case BooleanType.TRUE : mxType.setNullable(true); break;
-			case BooleanType.FALSE : mxType.setNullable(false); break;
+			case BooleanType.TRUE : mixedType.setNullable(true); break;
+			case BooleanType.FALSE : mixedType.setNullable(false); break;
 			default : 
 				throw new SchemaException(nullable, String.format(ATTRIBUTE_INVALID, 
 					Attribute.NULLABLE.tag, nullable.getValue(), "must be 'true' or 'false'"));
@@ -436,18 +436,18 @@ public final class SDSParser implements Parser {
 		// Set the pattern (not allowed on the any type).
 		Node pattern = getAttribute(sds, Attribute.PATTERN, isAnyType? null : false);
 		try { 
-			if (pattern != null) mxType.setPatternExpr(pattern.getValue()); 
+			if (pattern != null) mixedType.setPatternExpr(pattern.getValue()); 
 		} catch (PatternSyntaxException e) {
 			throw new SchemaException(pattern, 
 				String.format(ATTRIBUTE_INVALID, Attribute.PATTERN.tag, pattern.getValue(), e.getMessage()));
 		}
 		
 		// Set the length (only allowed on string and binary types).
-		Node length = getAttribute(sds, Attribute.LENGTH, mxType instanceof AbstractStringType ? false : null);
+		Node length = getAttribute(sds, Attribute.LENGTH, mixedType instanceof AbstractStringType ? false : null);
 		if (length != null) {
 			try {
 				NaturalInterval interval = NaturalInterval.from(length.getValue());
-				((AbstractStringType) mxType).setLength(interval);
+				((AbstractStringType) mixedType).setLength(interval);
 			} catch (IllegalArgumentException e) {
 				throw new SchemaException(length, String.format(ATTRIBUTE_INVALID, 
 					Attribute.LENGTH.tag, length.getValue(), e.getMessage()));
@@ -455,7 +455,7 @@ public final class SDSParser implements Parser {
 		}
 		
 		// Set the value range (only allowed on ranged types).
-		Node range = getAttribute(sds, Attribute.VALUE, mxType instanceof RangedType ? false : null);
+		Node range = getAttribute(sds, Attribute.VALUE, mixedType instanceof RangedType ? false : null);
 		if (range != null) {
 			try {	
 				Interval<T> interval;
@@ -467,14 +467,14 @@ public final class SDSParser implements Parser {
 					default: // we will never get here, unless we forgot to implement something
 						throw new RuntimeException("SDS type '" + content + "' not implemented!");
 				}
-				((RangedType<Comparable<?>>) mxType).setRange((Interval<T>) interval);
+				((RangedType<Comparable<?>>) mixedType).setRange((Interval<T>) interval);
 			} catch (IllegalArgumentException e) {
 				throw new SchemaException(range, 
 					String.format(ATTRIBUTE_INVALID, Attribute.VALUE.tag, range.getValue(), e.getMessage()));
 			}
 		}
 		
-		return mxType;
+		return mixedType;
 	}